+++ pkg/inst/NEWS.md
@@ -2,6 +2,7 @@
 * Prediction intervals are now created for `nnetar` objects in the ensemble. This should address one aspect of incorrect prediction intervals (e.g. issue #37).
 * theta models can be added (by including "`f`" in the `models=` argument for `hybridModel()`)
 * `accuracy.cvts` is exported
+* Time series must be at least four observations long
 
 # Version 0.2.0 [2016-09-23]
 * Add timeseries cross validation with `cvts()`